desafio3/api/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1
werkzeug==3.0.1